requirements.txt
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
phonenumberslite==8.13.44
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
